refactor: 将 Tauri 入口从 lib 移回 main
